requirements.txt
Flask
pygithub
